# config/openapi.yml
openapi: 3.1.0
info:
  title: Umoja Afya Enterprise EHR
  description: Umoja Afya Tanzania enterprise electronic health record. Includes national
    MPI and patient registration, scheduling/referrals, ADT and unit management, longitudinal
    documentation, orders/results, flowsheets/eMAR, pharmacy, revenue cycle, supply
    chain, quality, public health, remote care, FHIR and audit. Patient-facing portal
    delivery is intentionally deferred from this release.
  version: 11.0.0
paths:
...
  /api/v1/patients:
    get:
      tags:
      - Patients
      summary: List Patients
      operationId: list_patients_api_v1_patients_get
      parameters:
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
          - type: 'null'
          title: Search
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{mpi_id}:
    get:
      tags:
      - Patients
      summary: Get Patient
      operationId: get_patient_api_v1_patients__mpi_id__get
      parameters:
      - name: mpi_id
        in: path
        required: true
        schema:
          type: string
          title: Mpi Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Patients
      summary: Update Patient
      operationId: update_patient_api_v1_patients__mpi_id__patch
      parameters:
      - name: mpi_id
        in: path
        required: true
        schema:
          type: string
          title: Mpi Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientUpdateIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/tracker:
    get:
      tags:
      - Patient Flow and ADT
      summary: Get Tracker
      operationId: get_tracker_api_v1_tracker_get
      parameters:
      - name: facility_code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Facility Code
      - name: include_discharged
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Discharged
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/encounters/{encounter_id}/status:
    patch:
      tags:
      - Patient Flow and ADT
      summary: Update Status
      operationId: update_status_api_v1_encounters__encounter_id__status_patch
      parameters:
      - name: encounter_id
        in: path
        required: true
        schema:
          type: string
          title: Encounter Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EncounterStatusUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/encounters/{encounter_id}/discharge:
    post:
      tags:
      - Patient Flow and ADT
      summary: Discharge
      operationId: discharge_api_v1_encounters__encounter_id__discharge_post
      parameters:
      - name: encounter_id
        in: path
        required: true
        schema:
          type: string
          title: Encounter Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DischargeIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/recent-discharges:
    get:
      tags:
      - Patient Flow and ADT
      summary: Recent Discharges
      operationId: recent_discharges_api_v1_recent_discharges_get
      parameters:
      - name: facility_code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Facility Code
      - name: hours
        in: query
        required: false
        schema:
          type: integer
          maximum: 720
          minimum: 1
          default: 72
          title: Hours
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DischargeIn:
      properties:
        disposition:
          type: string
          title: Disposition
        summary:
          type: string
          title: Summary
        follow_up:
          anyOf:
          - type: string
          - type: 'null'
          title: Follow Up
        actor:
          type: string
          title: Actor
          default: demo-provider
      type: object
      required:
      - disposition
      - summary
      title: DischargeIn
...
    EncounterStatus:
      type: string
      enum:
      - PRE_REGISTERED
      - ARRIVED
      - WAITING_REGISTRATION
      - REGISTERED
      - WAITING_TRIAGE
      - TRIAGED
      - READY_FOR_PROVIDER
      - ROOMED
      - IN_PROGRESS
      - WAITING_RESULTS
      - READY_FOR_DISCHARGE
      - DISCHARGED
      - TRANSFERRED
      - LEFT_WITHOUT_BEING_SEEN
      title: EncounterStatus
    EncounterStatusUpdate:
      properties:
        status:
          $ref: '#/components/schemas/EncounterStatus'
        location:
          anyOf:
          - type: string
          - type: 'null'
          title: Location
        room:
          anyOf:
          - type: string
          - type: 'null'
          title: Room
        provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
        acuity:
          anyOf:
          - type: string
          - type: 'null'
          title: Acuity
        actor:
          type: string
          title: Actor
          default: demo-user
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - status
      title: EncounterStatusUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PatientUpdateIn:
      properties:
        first_name:
          anyOf:
          - type: string
            maxLength: 120
            minLength: 1
          - type: 'null'
          title: First Name
        middle_name:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Middle Name
        last_name:
          anyOf:
          - type: string
            maxLength: 120
            minLength: 1
          - type: 'null'
          title: Last Name
        date_of_birth:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date Of Birth
        sex:
          anyOf:
          - type: string
            maxLength: 40
          - type: 'null'
          title: Sex
        phone:
          anyOf:
          - type: string
            maxLength: 80
          - type: 'null'
          title: Phone
        nida_number:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Nida Number
        address:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Address
        region:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Region
        district:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: District
        next_of_kin:
          anyOf:
          - type: string
            maxLength: 300
          - type: 'null'
          title: Next Of Kin
        payer:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Payer
        member_number:
          anyOf:
          - type: string
            maxLength: 180
          - type: 'null'
          title: Member Number
        consent_status:
          anyOf:
          - type: string
            maxLength: 80
          - type: 'null'
          title: Consent Status
        actor:
          type: string
          maxLength: 180
          minLength: 2
          title: Actor
          default: Registration User
      type: object
      title: PatientUpdateIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...